requirements.txt
streamlit
numpy
pillow
pyautogui pillow
torch